feat web create DialogContext
Using context,provider and react protal to create Dialog component.
Before the Dialog state is managed by redux and now remove it, cuz
it will easy to duplicate the Dialog component. So I find the easy
way that using context to wrap the app and use protal to make the
dialog outside the app in html.